Rewrite of pre-detection tests (#780)
* refactored pre-dection tests
* added changelog entry
* reworked stop_logprep for safe process termination


* added structured yaml file

---------

+++ tests/acceptance/util.py
@@ -10,6 +10,7 @@ import re
 import socketserver
 import subprocess
 import sys
+import tempfile
 import threading
 import time
 from copy import deepcopy
@@ -229,8 +230,8 @@ def get_default_logprep_config(pipeline_config, with_hmac=True) -> Configuration
         "output": {
             "jsonl": {
                 "type": "jsonl_output",
-                "output_file": "tests/testdata/acceptance/test_kafka_data_processing_acceptance.out",
-                "output_file_custom": "tests/testdata/acceptance/test_kafka_data_processing_acceptance_custom.out",
+                "output_file": tempfile.mkstemp(suffix="output.jsonl")[1],
+                "output_file_custom": tempfile.mkstemp(suffix="custom.jsonl")[1],
             }
         },
     }
@@ -288,10 +289,28 @@ def wait_for_output(
 
 
 def stop_logprep(proc: subprocess.Popen) -> None:
-    process = psutil.Process(proc.pid)
-    for p in process.children(recursive=True):
-        p.kill()
-    process.kill()
+    if proc is None or not psutil.pid_exists(proc.pid):
+        return
+
+    try:
+        process = psutil.Process(proc.pid)
+        for p in process.children(recursive=True):
+            if p.is_running():
+                p.terminate()
+
+        process.wait(timeout=5)
+        process.terminate()
+        for p in process.children(recursive=True):
+            if p.is_running():
+                p.kill()
+
+        if process.is_running():
+            process.kill()
+
+    except (psutil.NoSuchProcess, psutil.ZombieProcess):
+        pass
+    except psutil.TimeoutExpired:
+        process.kill()
 
 
 def get_full_pipeline(exclude=None):
